Separate UpdateChapter component
